refactor keymaps
