+++ lua/plugins/colorscheme/catppuccin.lua
