+++ lua/plugins/colorscheme/tokyonight.lua
